Remove redundant `x as T` casts flagged by clippy::unnecessary_cast
Ran `cargo clippy --fix -- -W clippy::unnecessary_cast` crate-wide (1237
sites, auto-fixable machine suggestions only rewrite a literal's
representation, e.g. `0 as u32` -> `0_u32`) plus 26 manual fixes for
same-type pointer casts clippy's autofix left for a human (each verified
against the concrete type alias/struct field before removing, e.g.
`FontFilePointer = *mut u8`, `Tsi5Table = ClassDef`, `GlyphHandle = Handle`).

Also fixes 3 now-unused `Handle` imports uncovered by removing the casts
that referenced it by name, and rewrites one `<= 0_u32` comparison in
libcff/subr.rs to clippy's suggested `== 0_u32` (a separate, pre-existing
absurd_extreme_comparisons lint that the redundant cast had been
obscuring from clippy's own detector -- behaviorally identical for
unsigned wrapping arithmetic).

Drops the `unnecessary_cast` ratchet entry from Cargo.toml's ignore list
now that the count is 0.

+++ rust/src/consolidate/otl/gpos_cursive.rs
@@ -87,5 +87,5 @@ pub unsafe fn consolidate_gpos_cursive(
             exit,
         });
     }
-    return (*subtable).len() == 0 as usize;
+    return (*subtable).len() == 0_usize;
 }
